backend/requirements.txt
asgiref==3.9.1
Django==4.2.7
django-cors-headers==4.2.0
djangorestframework==3.14.0
python-dotenv==1.0.0
sqlparse==0.5.3
typing_extensions==4.14.1
